Change loss to optax huber loss, calculate loss using full size images not regions of interest, implement regularization
The aspect ratio parameter was collapsing to zero, and so was the WDH model image, so there is a bug in the way we are defining the region of interest. Regularize the beta, a_r, and sig params for cases when optimizing multiple WDH models such that any additional models will have parameters that do not deviate more than 10% of the dominant WDH model.